term: don't pack the outer struct, in the row struct
This appears to improve performance, back to the levels in the master
branch.

The struct is now 24 bytes, which is still a reduction of 25%, down
from 32 bytes.

Initial benchmarks suggest we're at the same performance levels as the
master branch on all vtetest tests, and actually somewhat faster on
some of the scroll tests.