term: shutdown: special case child terminating due to SIGHUP
When closing the window via a compositor shortcut, or clicking the CSD
close button, we kill the client application by sending it SIGHUP (and
then SIGTERM followed by SIGKILL, if it refuses to die).

Many shells either don't catch SIGHUP, causing us to exit with a
non-zero value, or they catch it, and then _they_ exit with a non-zero
value, causing _us_ to propagate that to the user.

This patch changes foot's behavior to "ignore" death-by-sighup, and
always exit with value 0.

The log message is changed to debug level, for SIGHUP only; other
death-by-signal is still logged at the warning level.