render: fix hollow cursor style for "extra" cursors
When the cursor is "block style", i.e. either when the window is
focused and the cursor is block, or when unfocused and we're supposed
to draw a hollow cursor, then we need to draw it *before* compositing
the glyph.

In all other cases (i.e. beam and underline cursors), it should be
rendered after the glyph.

With the addition of "extra" cursors, the logic for determining which
type it is has become much more complex, and the single if-statement
we had was wrong (it didn't handle "extra" cursors correctly).

Fix by splitting up all the checks into multiple if-statements.