render: dirty all multi-cursor cells *after* rendering a frame
This ensures they're re-rendered the next frame, and fixes an issue
were multi-cursors weren't erased when scrolled. That is, they
scrolled with the content, *and* was re-drawn at the correct (fixed)
location they were scrolled away from.

Since multi-cursors shouldn't scroll with the content, they need to be
erased when the content is scrolled.